Move existing controller to v1alpha2
