+++ pkg/controller/v1alpha2/controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package v1alpha2
 
 import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
